requirements.txt
# Core Framework
fastapi==0.135.1
uvicorn==0.41.0
Jinja2==3.1.6
python-multipart==0.0.22
starlette==0.52.1

# Database
SQLAlchemy==2.0.48
PyMySQL==1.1.1           # MySQL driver (PythonAnywhere production)

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
cryptography==46.0.5

# Pydantic + Email validation
pydantic==2.12.5
pydantic[email]
email-validator==2.3.0

# Utils
annotated-types==0.7.0
anyio==4.12.1
h11==0.16.0
idna==3.11
dnspython==2.8.0
typing_extensions==4.15.0
